signup page / signup functionality

+++ src/pages/SignupPage.jsx
@@ -1,52 +1,54 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useFormik } from 'formik';
 import * as yup from 'yup';
 import axios from 'axios';
-import { useNavigate, NavLink } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { Container, Card, Row, Col, Form, Button, FloatingLabel, Image } from 'react-bootstrap';
 import useAuth from '../hook/useAuth.js';
-import logImage from '../assets/images/hexlet_chat.jpg';
+import signImage from '../assets/images/hexlet_chat.jpg';
 
-function Loginpage() {
-  const { logIn } = useAuth();
+function SignupPage() {
   const navigate = useNavigate();
-  const [errors, setErrors] = useState([]);
-
+  const { logIn } = useAuth();
   const formik = useFormik({
     initialValues: {
       username: '',
       password: '',
+      confirmPassword: '',
     },
     validationSchema: yup.object({
       username: yup.string()
-        .max(15, 'Must be 15 characters or less')
+        .min(3, 'Minimum length is 3 letters')
+        .max(20, 'Maximum length is 3 letters')
         .required('Обязательное поле'),
       password: yup.string()
+        .min(6, 'Password must contain 6 letters or more')
+        .required('Обязательное поле'),
+      confirmPassword: yup.string()
+        .oneOf([yup.ref('password'), null], 'Passwords must match')
         .required('Обязательное поле'),
     }),
     onSubmit: (values) => {
-      axios.post('/api/v1/login', values)
+      axios.post('/api/v1/signup', values)
         .then((res) => {
           localStorage.setItem('userId', JSON.stringify(res.data));
-          setErrors([]);
           logIn();
           navigate('/');
-        })
-        .catch(() => setErrors(['Password/login is incorrect']));
+        });
     },
   });
   return (
     <Card style={{ width: '60rem', height: '30rem' }} className="shadow-sm">
-      <Container fluid style={{ height: '100%', padding: '0px 12px'}}>
-        <Row style={{ height: '85%' }} className="align-items-center">
+      <Container fluid style={{ height: '100%', padding: '0px 12px' }}>
+        <Row style={{ height: '100%' }} className="align-items-center">
           <Col className="text-center">
-            <Image roundedCircle="true" src={logImage} />
+            <Image roundedCircle="true" src={signImage} />
           </Col>
           <Col className="text-center">
-            <h1>Войти</h1>
+            <h1>Регистрация</h1>
             <Form onSubmit={formik.handleSubmit} noValidate>
               <Form.Group md="4" controlId="validationFormikUsername">
-                <FloatingLabel label="Ваш ник" className="mb-3">
+                <FloatingLabel label="Имя пользователя" className="mb-3">
                   <Form.Control
                     id="username"
                     name="username"
@@ -54,7 +56,7 @@ function Loginpage() {
                     onChange={formik.handleChange}
                     value={formik.values.username}
                     onBlur={formik.handleBlur}
-                    placeholder="Ваш ник"
+                    placeholder="Имя пользователя"
                     isInvalid={formik.touched.username && formik.errors.username}
                   />
                   <Form.Control.Feedback type="invalid">{formik.errors.username}</Form.Control.Feedback>
@@ -75,19 +77,25 @@ function Loginpage() {
                   <Form.Control.Feedback type="invalid">{formik.errors.password}</Form.Control.Feedback>
                 </FloatingLabel>
               </Form.Group>
+              <Form.Group md="4" controlId="validationFormikPassword">
+                <FloatingLabel label="Подтвердите пароль" className="mb-3">
+                  <Form.Control
+                    id="confirmPassword"
+                    name="confirmPassword"
+                    type="password"
+                    onChange={formik.handleChange}
+                    value={formik.values.confirmPassword}
+                    onBlur={formik.handleBlur}
+                    placeholder="Подтвердите пароль"
+                    isInvalid={formik.touched.confirmPassword && formik.errors.confirmPassword}
+                  />
+                  <Form.Control.Feedback type="invalid">{formik.errors.confirmPassword}</Form.Control.Feedback>
+                </FloatingLabel>
+              </Form.Group>
               <Form.Group className="d-grid gap-2">
-                <Button variant="outline-primary" size="md" type="submit">Войти</Button>
-                {/* {errors.length === 0 
-                  ? null
-                  : <Form.Control.Feedback>{errors.map((error) => error)}</Form.Control.Feedback>} */}
+                <Button variant="outline-primary" size="md" type="submit">Зарегистрироваться</Button>
               </Form.Group>
             </Form>
-          </Col>
-        </Row>
-        <Row style={{ height: '15%', backgroundColor: '#f7f7f7' }} className="justify-content-md-center align-items-center">
-          <Col md="auto">
-            Нет аккаунта?
-            <NavLink to="/signup">Регистрация</NavLink>
           </Col>
         </Row>
       </Container>
@@ -95,4 +103,4 @@ function Loginpage() {
   );
 }
 
-export default Loginpage;
+export default SignupPage;
